requirements.txt
kurigram
python-dotenv
apscheduler
pyyaml
